+++ packages/client/src/rtc/codecs.ts
@@ -1,13 +1,15 @@
 import { isReactNative } from '../helpers/platforms';
 import { removeCodec, setPreferredCodec } from '../helpers/sdp-munging';
+import { getLogger } from '../logger';
 
 export const getPreferredCodecs = (
   kind: 'audio' | 'video',
   preferredCodec: string,
   codecToRemove?: string,
 ): RTCRtpCodecCapability[] | undefined => {
+  const logger = getLogger(['codecs']);
   if (!('getCapabilities' in RTCRtpSender)) {
-    console.warn('RTCRtpSender.getCapabilities is not supported');
+    logger?.('warn', 'RTCRtpSender.getCapabilities is not supported');
     return;
   }
   const cap = RTCRtpSender.getCapabilities(kind);
@@ -17,7 +19,7 @@ export const getPreferredCodecs = (
   const unmatched: RTCRtpCodecCapability[] = [];
   cap.codecs.forEach((c) => {
     const codec = c.mimeType.toLowerCase();
-    console.log(c);
+    logger?.('debug', `Found supported codec: ${codec}`);
     const shouldRemoveCodec =
       codecToRemove && codec === `${kind}/${codecToRemove}`;
     if (shouldRemoveCodec) return;
@@ -39,7 +41,9 @@ export const getPreferredCodecs = (
     matched.push(c);
   });
 
-  return [...matched, ...partialMatched, ...unmatched];
+  const result = [...matched, ...partialMatched, ...unmatched];
+  logger?.('info', `Preffered codecs: `, result);
+  return result;
 };
 
 export const getGenericSdp = async (